Break out Cassandra daemon for test
Relocate import script for cassandra.yaml in integraiton tests
