--- integration-test-custom-format/src/test/resources/prepare_cassandra_yaml.sh
@@ -1,36 +1,0 @@
-#!/bin/sh
-#
-# Copyright 2018 Telefonaktiebolaget LM Ericsson
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# Use this script to apply tags in the cassandra.yaml file in the same directory
-# These tags are used by the embedded cassandra daemon in the integration tests.
-
-sed -i 's/# \(data_file_directories:\)/\1/' cassandra.yaml
-sed -i 's@^# \(.*\)/var/lib/cassandra/\(.*\)@\1###tmp###/cassandra/\2@' cassandra.yaml
-sed -i 's/^\(authenticator: \).*/\1com.ericsson.bss.cassandra.ecaudit.auth.AuditPasswordAuthenticator/' cassandra.yaml
-sed -i 's/^\(authorizer: \).*/\1com.ericsson.bss.cassandra.ecaudit.auth.AuditAuthorizer/' cassandra.yaml
-sed -i 's/^\(role_manager: \).*/\1com.ericsson.bss.cassandra.ecaudit.auth.AuditRoleManager/' cassandra.yaml
-sed -i 's/roles_validity_in_ms:.*/roles_validity_in_ms: 0/' cassandra.yaml
-sed -i 's/permissions_validity_in_ms:.*/permissions_validity_in_ms: 0/' cassandra.yaml
-sed -i 's/credentials_validity_in_ms:.*/credentials_validity_in_ms: 0/' cassandra.yaml
-sed -i 's/key_cache_size_in_mb:.*/key_cache_size_in_mb: 0/' cassandra.yaml
-sed -i 's/counter_cache_size_in_mb:.*/counter_cache_size_in_mb: 0/' cassandra.yaml
-sed -i 's/^storage_port:.*/storage_port: ###storage_port###/' cassandra.yaml
-sed -i 's/^ssl_storage_port:.*/ssl_storage_port: ###ssl_storage_port###/' cassandra.yaml
-sed -i 's/^native_transport_port:.*/native_transport_port: ###native_transport_port###/' cassandra.yaml
-sed -i 's/^rpc_port:.*/rpc_port: ###rpc_port###/' cassandra.yaml
-sed -i 's/^num_tokens:.*/num_tokens: 1/' cassandra.yaml
-sed -i 's/^enable_user_defined_functions:.*/enable_user_defined_functions: true/' cassandra.yaml
